Add Skip instructions; Change register types on instructions; Update README.md

+++ src/chip8/errors.rs
@@ -7,5 +7,5 @@ pub enum SubroutineError {
 
 #[derive(Debug)]
 pub enum RegisterError {
-    InvalidRegister(u8),
+    InvalidRegister(usize),
 }
